prevent whole info page from rerender on uptime change
