+++ src/pages/node/info/index.tsx
@@ -1,54 +1,37 @@
 import { useState, useEffect } from 'react';
 import styled from '@emotion/styled';
-import { useAppDispatch, useAppSelector } from '../../store';
+import { useAppDispatch, useAppSelector } from '../../../store';
 import { Link } from 'react-router-dom';
-import { copyStringToClipboard } from '../../utils/functions';
+import { copyStringToClipboard } from '../../../utils/functions';
 import { formatEther } from 'viem';
 
 // Mui
 import { Paper } from '@mui/material';
 
 // HOPR Components
-import Section from '../../future-hopr-lib-components/Section';
-import { actionsAsync } from '../../store/slices/node/actionsAsync';
-import { TableExtended } from '../../future-hopr-lib-components/Table/columed-data';
-import { SubpageTitle } from '../../components/SubpageTitle';
-import Tooltip from '../../future-hopr-lib-components/Tooltip/tooltip-fixed-width';
-import WithdrawModal from '../../components/Modal/node/WithdrawModal';
-import SmallActionButton from '../../future-hopr-lib-components/Button/SmallActionButton';
-import { ColorStatus } from '../../components/InfoBar/details';
-import ProgressBar from '../../future-hopr-lib-components/Progressbar';
+import Section from '../../../future-hopr-lib-components/Section';
+import { actionsAsync } from '../../../store/slices/node/actionsAsync';
+import { TableExtended } from '../../../future-hopr-lib-components/Table/columed-data';
+import { SubpageTitle } from '../../../components/SubpageTitle';
+import Tooltip from '../../../future-hopr-lib-components/Tooltip/tooltip-fixed-width';
+import WithdrawModal from '../../../components/Modal/node/WithdrawModal';
+import SmallActionButton from '../../../future-hopr-lib-components/Button/SmallActionButton';
+import { ColorStatus } from '../../../components/InfoBar/details';
+import ProgressBar from '../../../future-hopr-lib-components/Progressbar';
 
 //Icons
 import CopyIcon from '@mui/icons-material/ContentCopy';
 import LaunchIcon from '@mui/icons-material/Launch';
+
+//Info Components
+import NodeUptime from './node-uptime';
+
 
 const TdActionIcons = styled.td`
   display: flex;
   gap: 8px;
   align-items: center;
 `;
-
-
-function Row(props: {
-  tooltip: any,
-  title: any,
-  value: any,
-}) {
-  return(
-    <tr>
-      <th>
-        <Tooltip
-          title={props.tooltip}
-          notWide
-        >
-        <span>{props.title}</span>
-      </Tooltip>
-      </th>
-      <td>{props.value}</td>
-    </tr>
-  )
-}
 
 function InfoPage() {
   const dispatch = useAppDispatch();
@@ -71,29 +54,12 @@ function InfoPage() {
   const statisticsFetching = useAppSelector((store) => store.node.statistics.isFetching);
   const nodeStartedEpoch = useAppSelector((store) => store.node.metrics.data.parsed?.hopr_up?.data[0]);
   const nodeStartedTime = nodeStartedEpoch && typeof(nodeStartedEpoch) === 'number'? new Date(nodeStartedEpoch*1000).toJSON().replace('T', ' ').replace('Z', ' UTC') : '-';
-  const [nodeTimeUp, set_nodeTimeUp] = useState('-');
   const nodeSync = useAppSelector((store) => store.node.metrics.data.parsed?.hopr_indexer_sync_progress?.data[0]);
 
 
   useEffect(() => {
     fetchInfoData();
   }, [apiEndpoint, apiToken]);
-
-  useEffect(() => {
-    let interval:any;
-    if(nodeStartedEpoch && typeof(nodeStartedEpoch) === 'number') {
-      interval = setInterval(() => {
-        const nodeStartedEpochMs = Math.floor(nodeStartedEpoch * 1000);
-        const uptimeSec = Math.floor((Date.now() - nodeStartedEpochMs) / 1000);
-        const days = Math.floor(uptimeSec / 86400);
-        const hours = Math.floor((uptimeSec-(days * 86400)) / 3600);
-        const minutes = Math.floor((uptimeSec-(days * 86400)-(hours * 3600)) / 60);
-        const seconds = Math.floor((uptimeSec-(days * 86400)-(hours * 3600)-(minutes * 60)));
-        set_nodeTimeUp(`${days} days ${hours} hours ${minutes} min ${seconds} sec`)
-      }, 1_000)
-    }
-    return () => clearInterval(interval);
-  }, [nodeStartedEpoch]);
 
   const fetchInfoData = () => {
     if (!apiEndpoint || !apiToken) return;
@@ -581,7 +547,9 @@ function InfoPage() {
               </th>
               <td>{info?.network}</td>
             </tr>
-            <tr>
+            <tr
+              key='node-startdate'
+            >
               <th>
                 <Tooltip
                   title="Date when you node was started"
@@ -592,22 +560,7 @@ function InfoPage() {
               </th>
               <td>{nodeStartedTime}</td>
             </tr>
-            {/* <tr>
-              <th>
-                <Tooltip
-                  title="The amount of the node is up"
-                  notWide
-                >
-                  <span>Uptime</span>
-                </Tooltip>
-              </th>
-              <td>{nodeTimeUp}</td>
-            </tr> */}
-            <Row
-              title={"Uptime"}
-              tooltip={"The amount of the node is up"}
-              value={nodeTimeUp}
-            />
+            <NodeUptime />
           </tbody>
         </TableExtended>
