+++ src/router.tsx
@@ -110,13 +110,6 @@ export const applicationMapNode: ApplicationMapType = [
         element: <InfoPage />,
         loginNeeded: 'node',
       },
-      // {
-      //   name: 'LOGS',
-      //   path: 'logs',
-      //   icon: <TerminalIcon />,
-      //   element: <SectionLogs />,
-      //   loginNeeded: 'node',
-      // },
       {
         name: 'TICKETS',
         path: 'tickets',
